dark mode implemented / / tailwind css improvements

+++ weather-app/src/app/UI-components/CurrentForecast/index.tsx
@@ -2,6 +2,9 @@ import { useState, useEffect, useRef } from 'react';
 import { FaHeart } from 'react-icons/fa';
 import { Toast } from "primereact/toast"
 import { IoHeartDislike } from "react-icons/io5";
+import { useAppSelector } from '../../hooks';
+import classNames from 'classnames';
+import Swal from 'sweetalert2';
 
 interface CurrentForecastProps {
   currentForecast: any | {};
@@ -18,9 +21,10 @@ const daysOfWeek = [
   'Saturday',
 ];
 const MAX_LIKED = 5;
-let arr:any = []
 
 const CurrentForecast = ({ currentForecast, header }: CurrentForecastProps) => {
+  const theme = useAppSelector((state) => state.theme.theme) || "light"
+
   const [favoriteList, setFavoriteList] = useState<any[]>(
     JSON.parse(localStorage.getItem('likedPlaces') as string) || []
   );
@@ -39,6 +43,11 @@ const CurrentForecast = ({ currentForecast, header }: CurrentForecastProps) => {
         const updatedList = [...favoriteList, place];
         setFavoriteList(updatedList);
         localStorage.setItem('likedPlaces', JSON.stringify(updatedList));
+        toast?.current?.show({
+          severity: 'success',
+          summary: 'Added successfully',
+          detail: ` ${header} added to your favorite list.`,
+        });
       } else {
         if (toast.current) {
           toast.current.show({
@@ -50,10 +59,27 @@ const CurrentForecast = ({ currentForecast, header }: CurrentForecastProps) => {
       }
     } else {
       // remove of the liked place 
-      const updatedList = favoriteList.filter((p: any) => p.MobileLink !== placeIdentifier);
-      setIsShown(false)
-      setFavoriteList(updatedList);
-      localStorage.setItem('likedPlaces', JSON.stringify(updatedList));
+      return Swal.fire({
+        title: `Are you sure you want to remove ${header}?`,
+        // text: "",
+        icon: "warning",
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Confirm",
+        showCancelButton: true,
+        // cancelButtonText: "Cancel",
+    }).then((result) => {
+        if (!result.isConfirmed) {
+
+        } else {
+          const updatedList = favoriteList.filter((p: any) => p.MobileLink !== placeIdentifier);
+          setIsShown(false)
+          setFavoriteList(updatedList);
+          localStorage.setItem('likedPlaces', JSON.stringify(updatedList));
+        }
+
+    });
+     
     }
   };
 
@@ -61,10 +87,16 @@ const CurrentForecast = ({ currentForecast, header }: CurrentForecastProps) => {
     <div className="max-w-sm flex items-center justify-center">
       <Toast ref={toast} />
 
-      <div className="bg-white shadow-lg rounded-lg overflow-hidden w-full">
+      <div className={classNames({
+        " shadow-lg rounded-lg font-bold  overflow-hidden w-full":true,
+        "bg-white opacity-70 text-black": theme ==="light",
+        "bg-gray-500 opacity-90 text-white": theme ==="dark"
+
+
+    })}>
         <div className="px-4 py-5 sm:p-6">
           <div className="flex items-center justify-between">
-            <h3 className="text-lg font-medium text-gray-900">
+            <h3 className="text-2xl font-bold text-black">
               {header} Forecast
             </h3>
             <div className="flex items-center space-x-2">
@@ -73,25 +105,25 @@ const CurrentForecast = ({ currentForecast, header }: CurrentForecastProps) => {
                 className={`text-2xl transition-colors text-red-500`}
                 onClick={()=>handleToggleLike(currentForecast)}
               >
-           {!isShown   ?  <FaHeart className={`cursor-pointer text-red-500`} /> : <IoHeartDislike className={`cursor-pointer text-gray-500`} />}
+           {!isShown   ?  <FaHeart className={`cursor-pointer opacity-100  text-red-500`} /> : <IoHeartDislike className={`cursor-pointer `} />}
               </button>
             </div>
           </div>
           <div className="mt-5 flex-col flex items-start flex-wrap space-y-2">
-            <p className="flex flex-col items-start justify-start text-sm text-gray-500">
-              <span className="text-lg text-gray-500">
+            <p className="flex flex-col items-start justify-start text-sm ">
+              <span className="text-lg ">
                 <strong>{daysOfWeek[new Date(currentForecast.LocalObservationDateTime).getDay()]}</strong>
               </span>
               <br />
-              <span className="text-sm text-gray-500">
+              <span className="text-sm ">
                 <strong>Weather:</strong> {currentForecast.WeatherText}
               </span>
             </p>
-            <span className="text-sm text-gray-500">
+            <span className="text-sm ">
               <strong>Temp (F):</strong>{' '}
               {currentForecast.Temperature.Imperial.Value}°F
             </span>
-            <span className="text-sm text-gray-500">
+            <span className="text-sm ">
               <strong>Temp (C):</strong>{' '}
               {currentForecast.Temperature.Metric.Value}°C
             </span>
